requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
backtrader==1.9.78.123
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.47.0
frozenlist==1.4.1
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
